add logging for token
